Revert "Merge branch 'master' into feature_3"
This reverts commit 8484d8e793ae70e4237c6bf71251a6cd536044e6, reversing
changes made to 675f8636ab351b68791f38c91a391910c07ce56e.
