+++ README.md
@@ -92,76 +92,3 @@ $ ./gradlew run
 
 ### Report
 Когда всё будет готово, присылайте pull request с изменениями, результатами нагрузочного тестирования и профилирования, а также анализом результатов **по сравнению с предыдущей** (синхронной) версией.
-
-## Этап 4. Шардирование (deadline 2020-10-21 00:00:00 MSK)
-
-Реализуем горизонтальное масштабирование через поддержку **кластерных конфигураций**, состоящих из нескольких узлов, взаимодействующих друг с другом через реализованный HTTP API.
-Для этого в `ServiceFactory` передаётся статическая "топология", представленная в виде множества координат **всех** узлов кластера в формате `http://<host>:<port>`.
-
-`gradle run` теперь стартует `Cluster` из трёх нод.
-
-Кластер распределяет ключи между узлами **детерминированным образом**.
-В кластере хранится **только одна** копия данных.
-Нода, получившая запрос, **проксирует** его на узел, отвечающий за обслуживание соответствующего ключа.
-Таким образом, общая ёмкость кластера равна суммарной ёмкости входящих в него узлов.
-
-Реализуйте один из алгоритмов распределения данных между узлами, например, [consistent hashing](https://en.wikipedia.org/wiki/Consistent_hashing) или [rendezvous hashing](https://en.wikipedia.org/wiki/Rendezvous_hashing).
-
-### Report
-Присылайте pull request со своей реализацией поддержки кластерной конфигурации на review.
-Не забудьте нагрузить, отпрофилировать и проанализировать результаты профилирования под нагрузкой.
-С учётом шардирования набор тестов расширяется, поэтому не забывайте **подмёрдживать upstream**.
-
-## Этап 5. Репликация (deadline 2020-10-28 00:00:00 MSK)
-
-Реализуем поддержку хранения [нескольких реплик](https://en.wikipedia.org/wiki/Replication_(computing)) данных в кластере для обеспечения отказоустойчивости.
-
-HTTP API расширяется query-параметром `replicas`, содержащим количество узлов, которые должны подтвердить операцию, чтобы она считалась выполненной успешно.
-Значение параметра `replicas` указывается в формате `ack/from`, где:
-* `ack` -- сколько ответов нужно получить
-* `from` -- от какого количества узлов
-
-Таким образом, теперь узлы должны поддерживать расширенный протокол (совместимый с предыдущей версией):
-* HTTP `GET /v0/entity?id=<ID>[&replicas=ack/from]` -- получить данные по ключу `<ID>`. Возвращает:
-  * `200 OK` и данные, если ответили хотя бы `ack` из `from` реплик
-  * `404 Not Found`, если ни одна из `ack` реплик, вернувших ответ, не содержит данные (либо данные **удалены хотя бы** на одной из `ack` ответивших реплик)
-  * `504 Not Enough Replicas`, если не получили `200`/`404` от `ack` реплик из всего множества `from` реплик
-
-* HTTP `PUT /v0/entity?id=<ID>[&replicas=ack/from]` -- создать/перезаписать (upsert) данные по ключу `<ID>`. Возвращает:
-  * `201 Created`, если хотя бы `ack` из `from` реплик подтвердили операцию
-  * `504 Not Enough Replicas`, если не набралось `ack` подтверждений из всего множества `from` реплик
-
-* HTTP `DELETE /v0/entity?id=<ID>[&replicas=ack/from]` -- удалить данные по ключу `<ID>`. Возвращает:
-  * `202 Accepted`, если хотя бы `ack` из `from` реплик подтвердили операцию
-  * `504 Not Enough Replicas`, если не набралось `ack` подтверждений из всего множества `from` реплик
-
-Если параметр `replicas` не указан, то в качестве `ack` используется значение по умолчанию, равное **кворуму** от количества узлов в кластере,
-а `from` равен общему количеству узлов в кластере, например:
-* `1/1` для кластера из одного узла
-* `2/2` для кластера из двух узлов
-* `2/3` для кластера из трёх узлов
-* `3/4` для кластера из четырёх узлов
-* `3/5` для кластера из пяти узлов
-
-Выбор узлов-реплик (множества `from`) для каждого `<ID>` является **детерминированным**:
-* Множество узлов-реплик для фиксированного ID и меньшего значения `from` является строгим подмножеством для большего значения `from`
-* При `PUT` не сохраняется больше копий данных, чем указано в `from` (т.е. не стоит писать лишние копии данных на все реплики)
-
-Фактически, с помощью параметра `replicas` клиент выбирает, сколько копий данных он хочет хранить, а также
-уровень консистентности при выполнении последовательности операций для одного ID.
-
-Таким образом, обеспечиваются следующие примеры инвариантов (список не исчерпывающий):
-* `GET` с `1/2` всегда вернёт данные, сохранённые с помощью `PUT` с `2/2` (даже при недоступности одной реплики при `GET`)
-* `GET` с `2/3` всегда вернёт данные, сохранённые с помощью `PUT` с `2/3` (даже при недоступности одной реплики при `GET`)
-* `GET` с `1/2` "увидит" результат `DELETE` с `2/2` (даже при недоступности одной реплики при `GET`)
-* `GET` с `2/3` "увидит" результат `DELETE` с `2/3` (даже при недоступности одной реплики при `GET`)
-* `GET` с `1/2` может не "увидеть" результат `PUT` с `1/2`
-* `GET` с `1/3` может не "увидеть" результат `PUT` с `2/3`
-* `GET` с `1/2` может вернуть данные несмотря на предшествующий `DELETE` с `1/2`
-* `GET` с `1/3` может вернуть данные несмотря на предшествующий `DELETE` с `2/3`
-* `GET` с `ack` равным `quorum(from)` "увидит" результат `PUT`/`DELETE` с `ack` равным `quorum(from)` даже при недоступности **<** `quorum(from)` реплик
-
-### Report
-Присылайте pull request со своей реализацией поддержки кластерной конфигурации на review.
-Не забудьте нагрузить, отпрофилировать и проанализировать результаты профилирования под нагрузкой.
-С учётом репликации набор тестов расширяется, поэтому не забывайте **подмёрдживать upstream**.
