+++ src/main/java/ru/mail/polis/service/ServiceFactory.java
@@ -21,7 +21,6 @@ import ru.mail.polis.dao.DAO;
 import ru.mail.polis.service.ivlev.ServiceImpl;
 
 import java.io.IOException;
-import java.util.Set;
 
 /**
  * Constructs {@link Service} instances.
@@ -38,16 +37,14 @@ public final class ServiceFactory {
     /**
      * Construct a storage instance.
      *
-     * @param port     port to bind HTTP server to
-     * @param dao      DAO to store the data
-     * @param topology a list of all cluster endpoints {@code http://<host>:<port>} (including this one)
+     * @param port port to bind HTTP server to
+     * @param dao  DAO to store the data
      * @return a storage instance
      */
     @NotNull
     public static Service create(
             final int port,
-            @NotNull final DAO dao,
-            @NotNull final Set<String> topology) throws IOException {
+            @NotNull final DAO dao) throws IOException {
         if (Runtime.getRuntime().maxMemory() > MAX_HEAP) {
             throw new IllegalStateException("The heap is too big. Consider setting Xmx.");
         }
