+++ src/test/java/ru/mail/polis/TestBase.java
@@ -87,9 +87,4 @@ abstract public class TestBase {
         result.rewind();
         return result;
     }
-
-    @NotNull
-    protected static String endpoint(final int port) {
-        return "http://localhost:" + port;
-    }
 }
