+++ src/test/java/ru/mail/polis/service/SingleNodeTest.java
@@ -31,7 +31,6 @@ import ru.mail.polis.dao.DAOFactory;
 import java.io.File;
 import java.io.IOException;
 import java.time.Duration;
-import java.util.Collections;
 import java.util.concurrent.TimeUnit;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -46,7 +45,6 @@ class SingleNodeTest extends TestBase {
     private static File data;
     private static DAO dao;
     private static int port;
-    private static String endpoint;
     private static Service storage;
     private static HttpClient client;
 
@@ -55,8 +53,7 @@ class SingleNodeTest extends TestBase {
         port = randomPort();
         data = Files.createTempDirectory();
         dao = DAOFactory.create(data);
-        endpoint = endpoint(port);
-        storage = ServiceFactory.create(port, dao, Collections.singleton(endpoint));
+        storage = ServiceFactory.create(port, dao);
         storage.start();
         Thread.sleep(TimeUnit.SECONDS.toMillis(1));
         reset();
@@ -245,8 +242,7 @@ class SingleNodeTest extends TestBase {
             java.nio.file.Files.createDirectory(data.toPath());
             dao = DAOFactory.create(data);
             port = randomPort();
-            endpoint = endpoint(port);
-            storage = ServiceFactory.create(port, dao, Collections.singleton(endpoint));
+            storage = ServiceFactory.create(port, dao);
             storage.start();
             Thread.sleep(TimeUnit.SECONDS.toMillis(1));
             reset();
